Fixed an issue during recording of a component's execution where "missing or modified key in exec_properties" was raised from MLMD when `exec_properties` both omitted an existing property and added a new property.
Previously, the execution type registration logic was inconsistent in that it allowed subsequent runs to succeed when there were (1) omitted execution properties (reusing the existing type), (2) new execution properties (adding to the existing type), but failed in the case (3) when there both omitted and new execution properties. Adds a test case that fails under the previous logic.

PiperOrigin-RevId: 340998209